cli-lib: Add explicit pollint cast for sign conv
Silences -Wsign-conversion

+++ src/prg/c/src/cli-lib/libswocclient/optionproc.c
@@ -8,7 +8,7 @@
  * Released under the GPLv3 only.\n
  * SPDX-License-Identifier: GPL-3.0
  *
- * @version _v1.1.10 ==== 18/05/2019_
+ * @version _v1.1.11 ==== 02/06/2019_
  */
 
 /* **********************************************************************
@@ -52,6 +52,9 @@
  *				type bool.				*
  * 07/09/2018	MG	1.1.9	Use new mgemessage struct initialiser.	*
  * 18/05/2019	MG	1.1.10	Merge sub-projects into one.		*
+ * 02/06/2019	MG	1.1.11	Add explicit cast to pollint for sign	*
+ *				conversion. It has already been		*
+ *				validated > 0.				*
  *									*
  ************************************************************************
  */
@@ -410,7 +413,7 @@ int swc_client_wait(char *cnumlocks)
 	if ((prg_err = swcom_validate_config()))
 		return prg_err;
 	do {
-		sleep(pollint);
+		sleep((unsigned int)pollint);
 		if ((prg_err = exch_msg(outgoing_msg, om_length, msg)))
 			return prg_err;
